% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamental-functions.R
\name{call_data_ibb}
\alias{call_data_ibb}
\title{Fundamental Call Function}
\usage{
call_data_ibb(
  call_phrase = "",
  root_url = "https://data.ibb.gov.tr/api/3/action/"
)
}
\arguments{
\item{call_phrase}{Suffix of the call url.}

\item{root_url}{Root of the calls.}
}
\description{
Fundamental call function of (almost) all IBB Open Data Portal (exceptions apply)
}
