% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{get_all_data_sources}
\alias{get_all_data_sources}
\title{Get All Data Sources}
\usage{
get_all_data_sources(get_all_info = FALSE)
}
\arguments{
\item{get_all_info}{If TRUE return the whole list, otherwise just return the required result data frame.}
}
\description{
Get All Data Sources
}
