% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamental-functions.R
\name{get_data_frame_ibb}
\alias{get_data_frame_ibb}
\title{Get Data Frame from Resource ID Value}
\usage{
get_data_frame_ibb(resource_id, limit = "", return_json = FALSE)
}
\arguments{
\item{resource_id}{Resource ID of the wanted data set}

\item{limit}{Maximum number of records}

\item{return_json}{If TRUE, return a json string instead of a data frame.}
}
\description{
Given a data resource ID, this function returns the data frame.
}
