% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{get_detailed_info_on_data_source}
\alias{get_detailed_info_on_data_source}
\title{Get Info on a Data Source}
\usage{
get_detailed_info_on_data_source(package_id, get_all_info = FALSE)
}
\arguments{
\item{package_id}{Package ID taken from the package list.}

\item{get_all_info}{If TRUE return the whole list, otherwise just return the required result data frame.}
}
\description{
Get Info on a Data Source
}
