% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-calls.R
\name{get_environmental_monitoring_and_control_data}
\alias{get_environmental_monitoring_and_control_data}
\title{Environmental Monitoring and Control Data Based on Location in 2020}
\usage{
get_environmental_monitoring_and_control_data(
  params = list(limit = Sys.getenv("IBB_QUERY_LIMIT"))
)
}
\arguments{
\item{params}{List of parameters}
}
\description{
It contains environmental monitoring and control data with 13 parameters, measured in 15-minute periods in 38 pre-determined locations in Istanbul for 2020.
}
\examples{
\dontrun{
get_environmental_monitoring_and_control_data(params = list(limit = 5))
}
}
