% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-calls.R
\name{get_ibb_wifi_daily_data_usage}
\alias{get_ibb_wifi_daily_data_usage}
\title{ibbWiFi Daily Data Usage}
\usage{
get_ibb_wifi_daily_data_usage(params = list(limit =
  Sys.getenv("IBB_QUERY_LIMIT")))
}
\arguments{
\item{params}{List of parameters}
}
\description{
Data set contains info about number of locations, number of subscribers, number of sessions, number of devices, data usage and usage duration.
}
