% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-calls.R
\name{get_ibb_wifi_daily_new_subscribers}
\alias{get_ibb_wifi_daily_new_subscribers}
\title{ibbWiFi Daily New Subscribers}
\usage{
get_ibb_wifi_daily_new_subscribers(params = list(limit =
  Sys.getenv("IBB_QUERY_LIMIT")))
}
\arguments{
\item{params}{List of parameters}
}
\description{
Data set of new subscribers
}
