% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-calls.R
\name{get_ibb_wifi_new_subscribers_by_location}
\alias{get_ibb_wifi_new_subscribers_by_location}
\title{ibbWiFi New Subscribers by Location}
\usage{
get_ibb_wifi_new_subscribers_by_location(params = list(limit =
  Sys.getenv("IBB_QUERY_LIMIT")))
}
\arguments{
\item{params}{List of parameters}
}
\description{
Data set of new subscribers by subscription location
}
