% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-calls.R
\name{get_iski_call_center_complaint_stats}
\alias{get_iski_call_center_complaint_stats}
\title{Water Outage Complaints and Response Numbers by District in 2019}
\usage{
get_iski_call_center_complaint_stats(
  params = list(limit = Sys.getenv("IBB_QUERY_LIMIT"))
)
}
\arguments{
\item{params}{List of parameters}
}
\description{
Contains information on the number of complaints and responses to water outages and responses from all districts and neighborhoods across Istanbul in 2019, from the Istanbul Water and Sewerage Administration (İSKİ) call center.
}
\examples{
\dontrun{
get_iski_call_center_complaint_stats(params = list(limit = 5))
}
}
