% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-calls.R
\name{get_ispark_locations}
\alias{get_ispark_locations}
\title{Get ISPARK Locations}
\usage{
get_ispark_locations(params = list(limit = Sys.getenv("IBB_QUERY_LIMIT")))
}
\arguments{
\item{params}{List of parameters}
}
\description{
Get ISPARK Locations
}
\examples{
\dontrun{
get_ispark_locations(params = list(limit = 5))
}
}
