% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-calls.R
\name{get_park_garden_green_area_data}
\alias{get_park_garden_green_area_data}
\title{Park, Garden and Green Area Data for 2019}
\usage{
get_park_garden_green_area_data(params = list(limit =
  Sys.getenv("IBB_QUERY_LIMIT")))
}
\arguments{
\item{params}{List of parameters}
}
\description{
Includes the number of parks, number of trees maintained, number of sports grounds, length of sports fields, number of playgrounds, number of trees planted, and type of green areas in Istanbul in 2019.
}
\examples{
\dontrun{
get_park_garden_green_area_data(params=list(limit=5))
}
}
