% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibb-package.R
\docType{package}
\name{ibb-package}
\alias{ibb}
\alias{ibb-package}
\title{ibb: R Wrapper for Istanbul Municipality Open Data Portal}
\description{
Call wrappers for Istanbul Metropolitan
    Municipality's Open Data Portal (Turkish: İstanbul Büyükşehir
    Belediyesi Açık Veri Portalı) at <https://data.ibb.gov.tr/en/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/berkorbay/ibb}
  \item Report bugs at \url{https://github.com/berkorbay/ibb/issues}
}

}
\author{
\strong{Maintainer}: Berk Orbay \email{berk.orbay@algopoly.com}

}
\keyword{internal}
