% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{ibb_data_sources}
\alias{ibb_data_sources}
\title{List of Data Sources}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 205 rows and 23 columns.}
\usage{
ibb_data_sources
}
\description{
Data frame containing all resource ids and data sources of IBB Open Data Portal
}
\keyword{datasets}
