% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-functions.R
\name{set_query_limit}
\alias{set_query_limit}
\title{Set Query Limit}
\usage{
set_query_limit(limit = 100)
}
\arguments{
\item{limit}{Limit of the maximum results}
}
\description{
Sets maximum number of returned results
}
